+++ covalent/_workflow/electron.py
@@ -238,7 +238,13 @@ class Electron:
 
                 get_item.__name__ = node_name
 
-                get_item_electron = Electron(function=get_item, metadata=self.metadata.copy())
+                iterable_metadata = self.metadata.copy()
+                iterable_metadata[
+                    "call_before"
+                ] = (
+                    []
+                )  # Including this metadata from the parent electron in the collected nodes breaks file transfer
+                get_item_electron = Electron(function=get_item, metadata=iterable_metadata)
                 yield get_item_electron(self, i)
 
     def __getattr__(self, attr: str) -> "Electron":